requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pytesseract>=0.3.10
Pillow>=10.2.0
python-dotenv>=1.0.0
openai>=1.3.5
websockets>=12.0
aiofiles>=23.2.1
requests>=2.31.0
# ortools>=9.8.3296  # Optional: Install manually if Python 3.8-3.12 is available
# Note: OR-Tools may not support Python 3.14. The system will use a fallback algorithm if not available.
geopy>=2.4.1
